requirements.txt
# Core dependencies
pyyaml>=6.0

# Environment variable management
python-dotenv>=1.0.0

# LLM API clients (2024+ versions)
google-genai>=0.1.0         # Google Gemini (new SDK)
openai>=1.0.0               # OpenAI GPT models
anthropic>=0.7.0            # Anthropic Claude models

# HTTP client (for API calls)
requests>=2.31.0

# Testing (optional)
pytest>=7.0.0
pytest-mock>=3.10.0
